+++ packages/examples/src/examples/BridgedDevicesNode.ts
@@ -17,7 +17,7 @@ import { BridgedDeviceBasicInformationServer } from "@matter/main/behaviors/brid
 import { OnOffLightDevice } from "@matter/main/devices/on-off-light";
 import { OnOffPlugInUnitDevice } from "@matter/main/devices/on-off-plug-in-unit";
 import { AggregatorEndpoint } from "@matter/main/endpoints/aggregator";
-import { logEndpoint } from "@matter/protocol";
+import { logEndpoint } from "@matter/main/protocol";
 import { execSync } from "child_process";
 
 /** Initialize configuration values */