+++ packages/examples/src/examples/cluster/DummyThreadNetworkCommissioningServer.ts
@@ -7,8 +7,8 @@
 import { Bytes, Logger } from "@matter/main";
 import { GeneralCommissioningBehavior } from "@matter/main/behaviors/general-commissioning";
 import { NetworkCommissioningBehavior } from "@matter/main/behaviors/network-commissioning";
+import { NetworkCommissioning } from "@matter/main/clusters";
 import { DeviceAdvertiser, DeviceCommissioner } from "@matter/main/protocol";
-import { NetworkCommissioning } from "@matter/types/clusters";
 
 const firstNetworkId = new Uint8Array(32);